requirements.txt
asgiref==3.2.10
backcall==0.2.0
backports.pbkdf2==0.1
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
Django==3.1.1
django-rest-framework-jet==0.0.4
djangorestframework==3.11.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
json-encrypted-token==0.0.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
